feat(ui): extract experiences to own file. Input real data
