+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Header from './components/Header';
 import About from './components/About';
-import Experience from './components/Experience';
+import Experience from './components/experience';
 import Skills from './components/Skills';
 import Education from './components/Education';
 import Projects from './components/Projects';
