+++ src/components/experience/experience.ts
@@ -1,0 +1,91 @@
+export const experiences = [
+    {
+      id: "e1",
+      title: "Senior Software Engineer",
+      company: "Marsh McLennan",
+      location: "New York, NY",
+      duration: "2025 - Present",
+      description: "Development Center of Excellence Team Project: Polaris Launchpad",
+      achievements: [
+        "Built MCP Client to integrate with our internal MCP server leveraging AI capabilities",
+        "Contributed continuous feature improvements to enterprise application starter kit",
+        "Collaborated with cross-functional teams based on user feedback",
+        "Point of contact for level III support issues"
+      ],
+      technologies: ["React", "TypeScript", "Angular", "NodeJS", "NX", "Jest","Cypress",".NET", "Python"],
+      highlight: true
+    },
+    {
+      id: "e2",
+      title: "Senior Software Engineer",
+      company: "Marsh McLennan",
+      location: "New York, NY",
+      duration: "2024 - Present",
+      description: "Development Center of Excellence Team Project: Developer Portal",
+      achievements: [
+        "Spearheaded the development of a centralized hub to enhance the developer experience and streamline discovery processes through advanced tool integration while upholding development center of excellence standards to ensure high-quality outputs",
+        "Crafted an intuitive user experience for developers utilizing the platform, significantly",
+      ],
+      technologies: ["React", "TypeScript", "Tailwind CSS", "NodeJS", "NX", "Jest","Cypress"],
+      highlight: true
+    },
+    {
+      id: "e3",
+      title: "Senior Software Engineer",
+      company: "Marsh McLennan",
+      location: "New York, NY",
+      duration: "2022-2024",
+      description: "Platform Engineering Team Project: MMC Tech Platform Portal Redesign",
+      achievements: [
+        "Led the redesign of a fully managed, open-source, Kubernetes-based shared service landing platform as a micro front-end application",
+        "Crafted an intuitive user experience for developers utilizing the platform, significantly reducing complexity by leveraging Platform Engineering RESTful APIs" ],
+      technologies: ["React", "Material UI", "Vite"],
+      highlight: false
+    },
+    {
+      id: "e4",
+      title: "Senior Software Engineer",
+      company: "Marsh McLennan",
+      location: "New York, NY",
+      duration: "2021-2021",
+      description: "Application Engineering Team Project: New Hire Technology Setup",
+      achievements: [
+        "Developed a data-driven onboarding system for new hires, implementing automation through the New Hire Technology Setup (NHTS) to enhance global efficiency with minimal human intervention",
+        "Ensured that all new hires received their technology on day one by automating default technology triggers if selections were not made by hiring manager seven days prior to their start date"
+      ],
+      technologies: ["BPMN automation", "SQL", "JavaScript", "SOAP APIs", "C#"],
+      highlight: false
+    },
+    {
+      id: "e5",
+      title: "Senior Software Engineer",
+      company: "Marsh McLennan",
+      location: "New York, NY",
+      duration: "2019-2020",
+      description: "Application Engineering Team Project: MMC Technology Onboarding",
+      achievements: [
+        "Executed a comprehensive initiative to streamline the adoption of new technologies into the technology catalog, enhancing global processes",
+        "This initiative facilitated better management of incoming technologies, ensuring compliance with the firm's technology and security standards through a collaborative global effort",
+        "Conducted analysis and design of development efforts, providing clear implementation steps for the team",
+        "Delivered Level II & III support for the self-service technology catalog and focused on workflow and process automation"
+      ],
+      technologies: ["BPMN automation", "SQL", "JavaScript", "SOAP APIs", "C#"],
+      highlight: false
+    }
+  ];
+
+  export const designExperience = [{
+    id:"de1",
+    title: "Graphic Designer & Production Manager",
+    company: "Various Publishing Companies",
+    location: "Seattle, WA",
+    duration: "2006 - 2018",
+    description: "12 years in magazine print production, developing expertise in typography, layout, color theory, and design systems.",
+    achievements: [
+      "Managed end-to-end production for 15+ monthly publications with circulations up to 100K",
+      "Developed brand guidelines and design systems for consistent visual identity",
+      "Mastered typography, color theory, and layout principles for print and digital media",
+      "Led design teams and collaborated with editors, photographers, and marketing departments"
+    ],
+    technologies: ["Adobe Creative Suite", "InDesign", "Photoshop", "Illustrator", "Print Production", "Brand Design"]
+  }];
